UPSTREAM: coresight: tmc-etf: Fix snapshot mode update function
When working in snapshot mode function perf_aux_output_begin()
does not set the handle->size because the size is expected to be
deduced by the placement of the "head" and "old" pointers in user
space.  As such there is no point in trying to adjust the amount
of data to copy to the ring buffer.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Reviewed-by: Leo Yan <leo.yan@linaro.org>
Reviewed-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Tested-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(Upstream commit 99f81eb9c51d499f65bac25597e9def22f6a32dc).

Bug: 140266694
Change-Id: I81fe21897251370cf7dd3178d67e0dcc5f206d6c
Signed-off-by: Yabin Cui <yabinc@google.com>